Use pickle for providing RAM_CONFIG env var
This is necessary because PyEPL uses Python 2 and stupidly doesn't
verify that strings that should be char * types in C are in fact
bytes rather than unicode strings. This caused crashes whenever
trying to record audio files since JSON uses UTF-8 encoding and
loads strings as strings (as it should) instead of as bytes.

Death to PyEPL.